improvements

+++ src/com/library/books/Journals.java
@@ -7,7 +7,7 @@ import java.util.Set;
 
 public class Journals extends Book {
 
-    public Journals(long bookId, Set<Person> authors, String title, double price, double edition) {
+    public Journals(long bookId, Set<Person> authors, String title, double price, int edition) {
         super(bookId, authors, title, price, edition);
     }
 }
